+++ chaoslib/types.py
@@ -1,11 +1,12 @@
 # -*- coding: utf-8 -*-
-from typing import Any, Dict, List, Tuple, Union
+import enum
+from typing import Any, Dict, List, NamedTuple, Tuple, Union
 
 __all__ = ["MicroservicesStatus", "Probe", "Action", "Experiment", "Layer",
            "TargetLayers", "Activity", "Journal", "Run", "Secrets", "Step",
            "Configuration", "Discovery", "DiscoveredActivities", "Extension",
            "DiscoveredSystemInfo", "Settings", "EventPayload", "Tolerance",
-           "Hypothesis", "Control"]
+           "Hypothesis", "Control", "Strategy", "Schedule"]
 
 
 Action = Dict[str, Any]
@@ -37,3 +38,35 @@ Tolerance = Union[int, str, bool, list, Dict[str, Any]]
 Extension = Dict[str, Any]
 Hypothesis = Dict[str, Any]
 Control = Dict[str, Any]
+
+
+class Strategy(enum.Enum):
+    BEFORE_METHOD = "before-method-only"
+    AFTER_METHOD = "after-method-only"
+    DURING_METHOD = "during-method-only"
+    DEFAULT = "default"
+    CONTINOUS = "continous"
+
+    @staticmethod
+    def from_string(value: str) -> 'Strategy':
+        if value == "default":
+            return Strategy.DEFAULT
+        elif value == "before-method-only":
+            return Strategy.BEFORE_METHOD
+        elif value == "after-method-only":
+            return Strategy.AFTER_METHOD
+        elif value == "during-method-only":
+            return Strategy.DURING_METHOD
+        elif value == "continously":
+            return Strategy.CONTINOUS
+        elif value == "continously-fail-fast":
+            return Strategy.CONTINOUS
+
+        raise ValueError("Unknown strategy")
+
+
+class Schedule:
+    def __init__(self, continous_hypothesis_frequency: float = 1.0,
+                 fail_fast: bool = False):
+        self.continous_hypothesis_frequency = continous_hypothesis_frequency
+        self.fail_fast = fail_fast